pyproject.toml
[project]
name = "display-thingy"
version = "0.1.0"
description = "E-paper display manager for Raspberry Pi with pluggable views"
requires-python = ">=3.12"
dependencies = [
    "pillow>=11.0",
    "httpx>=0.27",
    "pydantic-settings>=2.0",
]

[project.optional-dependencies]
pi = [
    "gpiozero",
    "spidev",
]
dev = [
    "ruff",
    "pytest",
]

[project.scripts]
display-thingy = "display_thingy.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/display_thingy"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
